PyCharmMiscProject/requirements.txt
fastapi
uvicorn
python-pptx
httpx